CognitiveServices/ComputerVision/requirements.txt
    azure-cognitiveservices-vision-computervision>=0.2.0
    